rename map.js to map_data.js and update references
