+++ index.html
@@ -15,7 +15,7 @@
     <link rel="stylesheet" href="diet.css">
     <script type="module" src="app.js"></script>
     <script type="module" src="cards.js"></script>
-    <script type="module" src="map.js"></script>
+    <script type="module" src="map_data.js"></script>
     <script src="https://cdn.jsdelivr.net/npm/chart.js"></script>
 </head>
